worker-service/requirements.txt
boto3
psycopg2-binary
redis
requests
